refactor: 노선별 추가 요금 기능 넣기 전 리팩터링
- token authorization 예외처리 방법 변경
- Exception 관련 클래스와 메서드를 패키지 통합
