+++ src/main/java/nextstep/subway/exception/AuthorizationException.java
@@ -1,4 +1,4 @@
-package nextstep.subway.auth.application;
+package nextstep.subway.exception;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
