+++ src/main/java/nextstep/subway/exception/ExceptionHandler.java
@@ -1,6 +1,5 @@
 package nextstep.subway.exception;
 
-import nextstep.subway.auth.application.InvalidTokenException;
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -22,6 +21,11 @@ public class ExceptionHandler {
     @org.springframework.web.bind.annotation.ExceptionHandler(RuntimeException.class)
     public ResponseEntity handleRuntimeException(RuntimeException e) {
         return ResponseEntity.badRequest().build();
+    }
+
+    @org.springframework.web.bind.annotation.ExceptionHandler(AuthorizationException.class)
+    public ResponseEntity handleAuthorizationException(AuthorizationException e) {
+        return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
     }
 
 }
